Add in database to staging executable
